+++ bencher/example/example_strings.py
@@ -1,15 +1,15 @@
-import bencher as bch
+import bencher as bn
 
 
-class TestPrinting(bch.ParametrizedSweep):
+class TestPrinting(bn.ParametrizedSweep):
     # INPUTS
-    a = bch.StringSweep(default=None, string_list=["a1", "a2"], allow_None=True)
-    b = bch.StringSweep(default=None, string_list=["b1", "b2"], allow_None=True)
-    c = bch.StringSweep(default=None, string_list=["c1", "c2"], allow_None=True)
-    d = bch.StringSweep(default=None, string_list=["d1", "d2"], allow_None=True)
+    a = bn.StringSweep(default=None, string_list=["a1", "a2"], allow_None=True)
+    b = bn.StringSweep(default=None, string_list=["b1", "b2"], allow_None=True)
+    c = bn.StringSweep(default=None, string_list=["c1", "c2"], allow_None=True)
+    d = bn.StringSweep(default=None, string_list=["d1", "d2"], allow_None=True)
 
     # RESULTS
-    result = bch.ResultString()
+    result = bn.ResultString()
 
     def __call__(self, **kwargs):
         self.update_params_from_kwargs(**kwargs)
@@ -23,12 +23,12 @@ class TestPrinting(bch.ParametrizedSweep):
             self.result += f",{self.d}"
         self.result += "\n\ttab\n\t\ttab2"
 
-        self.result = bch.tabs_in_markdown(self.result)
+        self.result = bn.tabs_in_markdown(self.result)
         return super().__call__()
 
 
-def example_strings(run_cfg: bch.BenchRunCfg | None = None) -> bch.Bench:
-    bench = bch.Bench("strings", TestPrinting(), run_cfg=run_cfg)
+def example_strings(run_cfg: bn.BenchRunCfg | None = None) -> bn.Bench:
+    bench = bn.Bench("strings", TestPrinting(), run_cfg=run_cfg)
 
     for s in [
         ["a"],
@@ -42,4 +42,4 @@ def example_strings(run_cfg: bch.BenchRunCfg | None = None) -> bch.Bench:
 
 
 if __name__ == "__main__":
-    bch.run(example_strings)
+    bn.run(example_strings)