+++ bencher/example/meta/example_meta_cat.py
@@ -1,15 +1,15 @@
-import bencher as bch
+import bencher as bn
 from bencher.example.meta.example_meta import BenchMeta
 
 
-def example_meta_cat(run_cfg: bch.BenchRunCfg | None = None) -> bch.Bench:
+def example_meta_cat(run_cfg: bn.BenchRunCfg | None = None) -> bn.Bench:
     bench = BenchMeta().to_bench(run_cfg)
 
     bench.plot_sweep(
         title="Sweeping Categorical Variables",
         input_vars=[
-            bch.p("categorical_vars", [1, 2, 3]),
-            bch.p("sample_with_repeats", [1, 2]),
+            bn.p("categorical_vars", [1, 2, 3]),
+            bn.p("sample_with_repeats", [1, 2]),
         ],
         const_vars=dict(float_vars=0),
     )
@@ -18,4 +18,4 @@ def example_meta_cat(run_cfg: bch.BenchRunCfg | None = None) -> bch.Bench:
 
 
 if __name__ == "__main__":
-    bch.run(example_meta_cat)
+    bn.run(example_meta_cat)